# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IT Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/TicketPriority"
        - name: assetId
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a new ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketWithComments"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTicket
      tags: [tickets]
      summary: Update a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/comments:
    post:
      operationId: addTicketComment
      tags: [tickets]
      summary: Add a comment to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketComment"
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: department
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/profile:
    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    UserRole:
      type: string
      enum:
        - general_user
        - support_staff
        - administrator
    AssetStatus:
      type: string
      enum:
        - available
        - assigned
        - in_maintenance
        - retired
        - lost
    AssetCategory:
      type: string
      enum:
        - laptop
        - desktop
        - monitor
        - phone
        - tablet
        - printer
        - server
        - networking
        - peripheral
        - other
    TicketStatus:
      type: string
      enum:
        - open
        - in_progress
        - on_hold
        - resolved
        - closed
    TicketPriority:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - role
        - createdAt
        - updatedAt
    Asset:
      type: object
      properties:
        id:
          type: string
        assetTag:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/AssetCategory"
        status:
          $ref: "#/components/schemas/AssetStatus"
        assignedTo:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - assetTag
        - name
        - category
        - status
        - createdAt
        - updatedAt
...
    TicketComment:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        commentText:
          type: string
        createdBy:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ticketId
        - commentText
        - createdBy
        - createdAt
    Ticket:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        asset:
          $ref: "#/components/schemas/Asset"
          nullable: true
        createdBy:
          $ref: "#/components/schemas/UserProfile"
        assignedTo:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
        - description
        - status
        - priority
        - createdBy
        - createdAt
        - updatedAt
    TicketWithComments:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/TicketComment"
          required:
            - comments
    TicketListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - data
        - total
        - page
        - limit
...
    CreateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TicketPriority"
        assetId:
          type: string
          nullable: true
      required:
        - title
        - description
        - priority
    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        assignedToId:
          type: string
          nullable: true
    CreateCommentRequest:
      type: object
      properties:
        commentText:
          type: string
      required:
        - commentText
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        department:
          type: string
          nullable: true