# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Asset } from "./asset";
import type { TicketPriority } from "./ticketPriority";
import type { TicketStatus } from "./ticketStatus";
import type { UserProfile } from "./userProfile";

export interface Ticket {
  id: string;
  title: string;
  description: string;
  status: TicketStatus;
  priority: TicketPriority;
  asset?: Asset | null;
  createdBy: UserProfile;
  assignedTo?: UserProfile | null;
  createdAt: Date;
  updatedAt: Date;
  closedAt?: Date | null;
}