# lib/api-zod/src/generated/types/ticketComment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from "./userProfile";

export interface TicketComment {
  id: string;
  ticketId: string;
  commentText: string;
  createdBy: UserProfile;
  createdAt: Date;
}