# lib/api-zod/src/generated/types/ticketStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  open: "open",
  in_progress: "in_progress",
  on_hold: "on_hold",
  resolved: "resolved",
  closed: "closed",
} as const;